+++ .github/docker-release.sh
@@ -2,8 +2,8 @@
 
 set -e
 
-SERVICES=( issue credit )
-VERSION=$(echo "$TRAVIS_TAG" | cut -c 2-)  # Remove v from version
+SERVICES=( issue credit basketserver )
+VERSION=$(echo "$SOURCE_TAG" | cut -c 2-)  # Remove v from version
 echo "Building and deploying docker images with version: $VERSION"
 
 for SERVICE in "${SERVICES[@]}"
@@ -14,15 +14,12 @@ do
 
   ./gradlew ":${SERVICE}:bootBuildImage"
 
-  docker tag "${IMAGE}:latest" "${IMAGE}:${VERSION}"
-
   echo "Uploading docker images for ${SERVICE}-service."
   # Login to dockerhubwith credentials
-  echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin
+  echo "$DOCKER_ACCESS_TOKEN" | docker login -u "$DOCKER_USERNAME" --password-stdin
 
   # Push docker image to dockerhub
   docker push "${IMAGE}:${VERSION}"
-  docker push "${IMAGE}:latest"
 
   echo "Finished deploying ${SERVICE}-service!"
 done