+++ basketserver/src/main/java/de/upb/crypto/incentive/basketserver/exceptions/BasketNotFoundException.java
@@ -1,0 +1,7 @@
+package de.upb.crypto.incentive.basketserver.exceptions;
+
+public class BasketNotFoundException extends BasketServiceException {
+    public BasketNotFoundException() {
+        super("Basket not Found");
+    }
+}