+++ deployment/deploy.sh
@@ -89,8 +89,10 @@ if [[ $LOCAL == true ]]; then
   ./gradlew clean build
   ./gradlew ":credit:bootBuildImage"
   ./gradlew ":issue:bootBuildImage"
+  ./gradlew ":basketserver:bootBuildImage"
   kind load docker-image upbcuk/incentive-service-issue --name kind-t2
   kind load docker-image upbcuk/incentive-service-credit --name kind-t2
+  kind load docker-image upbcuk/incentive-service-basketserver --name kind-t2
   set +e
 
   # Print available images
@@ -108,6 +110,7 @@ kubectl apply -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/mast
 header "Deploy services"
 kubectl apply -f deployment/credit-service.yaml
 kubectl apply -f deployment/issue-service.yaml
+kubectl apply -f deployment/basketserver-service.yaml
 
 echo ""
 big-sep