+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Next Release
+ - Add basic crypto-entities (public parameters, user/provider keys, token, ...) [#11](https://github.com/upbcuk/incentive-services/pull/11)
+
+## [0.4.0] - 2021-02-03
  - Add options for local deployment and deployment using the container images from dockerhub and fix ingress configuration in `deploy.sh` [#8](https://github.com/upbcuk/incentive-services/pull/8)
  - Fix wrong issue service configuration bug [#9](https://github.com/upbcuk/incentive-services/pull/9)
+ - Add basket server [#10](https://github.com/upbcuk/incentive-services/pull/10)
+ - Migrate from travis to github workflows and replace latest version by version file [#12](https://github.com/upbcuk/incentive-services/pull/12)
 
 ## [0.3.0] - 2021-01-12
  - Rename protocols to protocoldefinition and add cryptoprotocols module, include math and craco, add interfaces for cryptoprocools use [#6](https://github.com/upbcuk/incentive-services/pull/6)